+++ public/esc.js
@@ -1,0 +1,3 @@
+/** Escape text for HTML. */
+export const esc = (s) =>
+  String(s ?? '').replace(/[&<>"']/g, (c) => ({ '&': '&amp;', '<': '&lt;', '>': '&gt;', '"': '&quot;', "'": '&#39;' })[c]);